+++ src/showdown-katex.js
@@ -1,30 +1,31 @@
 /** !
-* @file showdown-katex: markdown + ( latex and/or asciimath )
-* @author [obedm503](https://github.com/obedm503/)
-* @git [git repo](https://github.com/obedm503/showdown-katex.git)
-* @examples https://obedm503.github.io/showdown-katex/
-* @license MIT
-*/
+ * @file showdown-katex: markdown + ( latex and/or asciimath )
+ * @author obedm503
+ * @git https://github.com/obedm503/showdown-katex.git
+ * @examples https://obedm503.github.io/showdown-katex/
+ * @license MIT
+ */
 import katex from 'katex';
-import renderMathInElement from 'katex/dist/contrib/auto-render.min'
+import renderMathInElement from 'katex/dist/contrib/auto-render';
 import asciimathToTex from './asciimath-to-tex';
 
 /**
- * @param {NodeListOf<Element>} elements
- * @param config
- * @param {boolean} isAsciimath
+ * @param {object} opts
+ * @param {NodeListOf<Element>} opts.elements
+ * @param opts.config
+ * @param {boolean} opts.isAsciimath
  */
-function renderBlockElements(elements, config, isAsciimath) {
+function renderBlockElements({ elements, config, isAsciimath }) {
   if (!elements.length) {
     return;
   }
-  for (let i = 0, len = elements.length; i < len; i++) {
-    const element = elements[i];
-    const input = element.innerHTML;
+
+  elements.forEach(element => {
+    const input = element.textContent;
     const latex = isAsciimath ? asciimathToTex(input) : input;
     const html = katex.renderToString(latex, config);
-    element.parentNode.outerHTML = `<span title="${ input }">${ html }</span>`;
-  }
+    element.parentNode.outerHTML = `<span title="${input.trim()}">${html}</span>`;
+  });
 }
 
 /**
@@ -33,7 +34,7 @@ function renderBlockElements(elements, config, isAsciimath) {
  * @returns {string} regexp escaped string
  */
 function escapeRegExp(str) {
-  return str.replace(/[-[\]/{}()*+?.\\$^|]/g, "\\$&");
+  return str.replace(/[-[\]/{}()*+?.\\$^|]/g, '\\$&');
 }
 
 // katex config
@@ -42,65 +43,72 @@ const getConfig = (config = {}) => ({
   throwOnError: false, // fail silently
   errorColor: '#ff0000',
   ...config,
-  delimiters: ([
-    { left: "$$", right: "$$", display: true },
-    { left: "\\[", right: "\\]", display: true },
-    { left: "\\(", right: "\\)", display: false },
+  delimiters: [
+    { left: '$$', right: '$$', display: false },
     { left: '~', right: '~', display: false, asciimath: true },
-    // TODO: this doesn't work because the && is escaped to &amp;&amp; and so it's not recognized
-    // 2019-01-08: it works now, don't know why
-    { left: '&&', right: '&&', display: true, asciimath: true },
-  ]).concat(config.delimiters || []),
+  ].concat(config.delimiters || []),
 });
 
-const showdownKatex = (userConfig) => () => {
+const showdownKatex = userConfig => () => {
+  const parser = new DOMParser();
+  const config = getConfig(userConfig);
+  const asciimathDelimiters = config.delimiters
+    .filter(item => item.asciimath)
+    .map(({ left, right }) => {
+      const test = new RegExp(
+        `${escapeRegExp(left)}(.*?)${escapeRegExp(right)}`,
+        'g',
+      );
+      const replacer = (match, asciimath) => {
+        return `${left}${asciimathToTex(asciimath)}${right}`;
+      };
+      return { test, replacer };
+    });
+
   return [
-    // escape regex characters and convert asciimath to tex
-    // the next sub-extension will convert tex to html
-    // TODO: ignore stuff inside code (backticks)
     {
       type: 'output',
-      filter(text = '') {
-        const config = getConfig(userConfig);
-        const delimiters = config.delimiters.filter(item => item.asciimath);
-        if (!delimiters.length) {
-          return text;
+      filter(html = '') {
+        // parse html
+        const wrapper = parser.parseFromString(html, 'text/html').body;
+
+        if (asciimathDelimiters.length) {
+          // convert inline asciimath to inline latex
+          // ignore anything in code and pre elements
+          wrapper.querySelectorAll(':not(code):not(pre)').forEach(el => {
+            /** @type Text[] */
+            const textNodes = [...el.childNodes].filter(
+              // skip "empty" text nodes
+              node => node.nodeName === '#text' && node.nodeValue.trim(),
+            );
+
+            textNodes.forEach(node => {
+              const newText = asciimathDelimiters.reduce(
+                (acc, { test, replacer }) => acc.replace(test, replacer),
+                node.nodeValue,
+              );
+              node.nodeValue = newText;
+            });
+          });
         }
 
-        return delimiters.reduce((acc, delimiter) => {
-          const test = new RegExp(
-            `${ escapeRegExp(delimiter.left) }(.*?)${ escapeRegExp(delimiter.right) }`,
-            'g',
-          );
-          return acc.replace(test, (match, asciimath) =>
-            `${ delimiter.left }${ asciimathToTex(asciimath) }${ delimiter.right }`
-          );
-        }, text);
-      },
-    },
-    // render math
-    {
-      type: 'output',
-      filter(html) {
-        const config = getConfig(userConfig);
-        // parse html inside a <div>
-        const div = document.createElement('div');
-        div.innerHTML = html;
+        // find the math in code blocks
+        const latex = wrapper.querySelectorAll('code.latex.language-latex');
+        const asciimath = wrapper.querySelectorAll(
+          'code.asciimath.language-asciimath',
+        );
 
-        // find our "code"
-        const latex = div.querySelectorAll('code.latex.language-latex');
-        const asciimath = div.querySelectorAll('code.asciimath.language-asciimath');
+        renderBlockElements({ elements: latex, config });
+        renderBlockElements({ elements: asciimath, config, isAsciimath: true });
 
-        renderBlockElements(latex, config);
-        renderBlockElements(asciimath, config, true);
-        renderMathInElement(div, config);
+        renderMathInElement(wrapper, config);
 
-        // return html without the initial <div>
-        return div.innerHTML;
+        // return html without the wrapper
+        return wrapper.innerHTML;
       },
     },
   ];
-}
+};
 
 // register extension with default config
 if (typeof window.showdown !== 'undefined') {